chore(trading): update buy thunks structure
