+++ suite-common/trading/src/index.ts
@@ -9,5 +9,5 @@ export type { TradingBuyState } from './reducers/buyReducer';
 export * from './middlewares/tradingMiddleware';
 export * from './actions/tradingActions';
 export * from './actions/buyActions';
-export * from './thunks/buy/buyThunks';
+export * from './thunks/buy';
 export * from './thunks/tradingThunks';
