+++ suite-common/trading/src/thunks/buy/__tests__/loadInfoThunk.test.ts
@@ -3,15 +3,15 @@ import { FiatCurrenciesProps } from 'invity-api';
 
 import { configureMockStore, extraDependenciesMock } from '@suite-common/test-utils';
 
+import { buyThunks } from '../';
 import { invityAPI } from '../../../invityAPI';
 import { buyInitialState } from '../../../reducers/__fixtures__/buyTradingReducer';
 import { prepareBuyReducer } from '../../../reducers/buyReducer';
 import { regional } from '../../../regional';
-import { buyThunks } from '../buyThunks';
 
 const buyTradingReducer = prepareBuyReducer(extraDependenciesMock);
 
-describe('Testing buy thunks', () => {
+describe('Testing loadInfoThunk', () => {
     jest.mock('../../../invityAPI');
 
     invityAPI.setInvityServersEnvironment = () => {};
@@ -35,7 +35,7 @@ describe('Testing buy thunks', () => {
         },
     });
 
-    it('testing loadInfoThunk - successful response', async () => {
+    it('successful response', async () => {
         const buyInfoAPI = {
             country: 'CZ',
             suggestedFiatCurrency: 'CZK',
@@ -62,7 +62,7 @@ describe('Testing buy thunks', () => {
         });
     });
 
-    it('testing loadInfoThunk - unsuccessful response', async () => {
+    it('unsuccessful response', async () => {
         invityAPI.getBuyList = () => Promise.resolve(undefined);
 
         const buyInfoData = await store.dispatch(buyThunks.loadInfoThunk()).unwrap();
