+++ suite-common/trading/src/thunks/buy/handleRequestThunk.ts
@@ -1,6 +1,6 @@
 import { MutableRefObject } from 'react';
 
-import { BuyProviderInfo, BuyTradeQuoteRequest, CryptoId, FiatCurrencyCode } from 'invity-api';
+import { BuyTradeQuoteRequest } from 'invity-api';
 
 import { createThunk } from '@suite-common/redux-utils';
 import { Network } from '@suite-common/wallet-config';
@@ -10,8 +10,6 @@ import { Timer } from '@trezor/react-utils';
 import { tradingBuyActions } from '../../actions/buyActions';
 import { tradingActions } from '../../actions/tradingActions';
 import { invityAPI } from '../../invityAPI';
-import { BuyInfo } from '../../reducers/buyReducer';
-import { regional } from '../../regional';
 import {
     cryptoIdToCoinSymbol,
     selectTradingBuyQuotesRequest,
@@ -26,47 +24,7 @@ import {
 } from '../../utils';
 import { buyUtils } from '../../utils/buy/buyUtils';
 
-const BUY_COMMON_PREFIX = '@trading-buy/thunk';
-
-const loadInfoThunk = createThunk<BuyInfo>(
-    `${BUY_COMMON_PREFIX}/loadInfo`,
-    async (_, { fulfillWithValue }) => {
-        const buyInfo = await invityAPI.getBuyList();
-
-        if (!buyInfo?.providers) {
-            return fulfillWithValue({
-                buyInfo: {
-                    country: regional.UNKNOWN_COUNTRY,
-                    providers: [],
-                    defaultAmountsOfFiatCurrencies: {} as Record<FiatCurrencyCode, number>,
-                },
-                providerInfos: {},
-                supportedFiatCurrencies: [],
-                supportedCryptoCurrencies: [],
-            });
-        }
-
-        const providerInfos: { [name: string]: BuyProviderInfo } = {};
-
-        buyInfo.providers.forEach(e => (providerInfos[e.name] = e));
-
-        const supportedFiatCurrencies: string[] = [];
-        const supportedCryptoCurrencies: CryptoId[] = [];
-        buyInfo.providers.forEach(provider => {
-            supportedFiatCurrencies.push(
-                ...provider.tradedFiatCurrencies.map(c => c.toLowerCase()),
-            );
-            supportedCryptoCurrencies.push(...provider.tradedCoins);
-        });
-
-        return fulfillWithValue({
-            buyInfo,
-            providerInfos,
-            supportedFiatCurrencies,
-            supportedCryptoCurrencies,
-        });
-    },
-);
+import { BUY_THUNK_COMMON_PREFIX } from './';
 
 type GetQuotesRequest = {
     requestData: BuyTradeQuoteRequest;
@@ -123,7 +81,7 @@ const getQuoteRequestData = ({
     return request;
 };
 
-export type HandleRequestThunk = {
+export type HandleRequestThunkProps = {
     formValues: TradingBuyFormProps;
     turnOffLoading: boolean;
     network: Network;
@@ -132,8 +90,8 @@ export type HandleRequestThunk = {
     abortControllerRef: MutableRefObject<AbortController | null>;
 };
 
-const handleRequestThunk = createThunk(
-    `${BUY_COMMON_PREFIX}/handleChange`,
+export const handleRequestThunk = createThunk(
+    `${BUY_THUNK_COMMON_PREFIX}/handleChange`,
     async (
         {
             formValues,
@@ -142,7 +100,7 @@ const handleRequestThunk = createThunk(
             timer,
             shouldSendInSats,
             abortControllerRef,
-        }: HandleRequestThunk,
+        }: HandleRequestThunkProps,
         { dispatch, getState },
     ) => {
         timer.loading();
@@ -202,8 +160,3 @@ const handleRequestThunk = createThunk(
         timer.reset();
     },
 );
-
-export const buyThunks = {
-    loadInfoThunk,
-    handleRequestThunk,
-};
