+++ suite-common/trading/src/thunks/buy/loadInfoThunk.ts
@@ -1,0 +1,49 @@
+import { BuyProviderInfo, CryptoId, FiatCurrencyCode } from 'invity-api';
+
+import { createThunk } from '@suite-common/redux-utils';
+
+import { invityAPI } from '../../invityAPI';
+import { BuyInfo } from '../../reducers/buyReducer';
+import { regional } from '../../regional';
+
+import { BUY_THUNK_COMMON_PREFIX } from './';
+
+export const loadInfoThunk = createThunk<BuyInfo>(
+    `${BUY_THUNK_COMMON_PREFIX}/loadInfo`,
+    async (_, { fulfillWithValue }) => {
+        const buyInfo = await invityAPI.getBuyList();
+
+        if (!buyInfo || !buyInfo.providers) {
+            return fulfillWithValue({
+                buyInfo: {
+                    country: regional.UNKNOWN_COUNTRY,
+                    providers: [],
+                    defaultAmountsOfFiatCurrencies: {} as Record<FiatCurrencyCode, number>,
+                },
+                providerInfos: {},
+                supportedFiatCurrencies: [],
+                supportedCryptoCurrencies: [],
+            });
+        }
+
+        const providerInfos: { [name: string]: BuyProviderInfo } = {};
+
+        buyInfo.providers.forEach(e => (providerInfos[e.name] = e));
+
+        const supportedFiatCurrencies: string[] = [];
+        const supportedCryptoCurrencies: CryptoId[] = [];
+        buyInfo.providers.forEach(provider => {
+            supportedFiatCurrencies.push(
+                ...provider.tradedFiatCurrencies.map(c => c.toLowerCase()),
+            );
+            supportedCryptoCurrencies.push(...provider.tradedCoins);
+        });
+
+        return fulfillWithValue({
+            buyInfo,
+            providerInfos,
+            supportedFiatCurrencies,
+            supportedCryptoCurrencies,
+        });
+    },
+);
